feat: 止まっている壁への衝突処理を追加
