+++ Assets/Scripts/StaticColliderWall.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e295737025ab82e5aa643e138570e771
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
